Move clas for handling superglobals to a dedicated folder + add  handling to Response
